backend/requirements.txt
django==3.2
black
drf-yasg
psycopg2-binary
djangorestframework-simplejwt
requests